requirements.txt
numpy
streamlit-sortables
scipy
scikit-learn
matplotlib
streamlit_ace
psutil